requirements.txt
asgiref==3.5.2
Django==4.0.5
django-appconf==1.0.5
django-imagekit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
pilkit==2.0
Pillow==9.1.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
psycopg2>=2.8